+++ modulo3/projeto-labex/labex/src/pages/AdminHomePage.js
@@ -29,38 +29,14 @@ export default function AdminHomePage() {
     useEffect(() => {
         getTrips()
     }, [])
-
-    const listaViagens = viagens.map((trip) => {
-        const irParaDetalhes = () => {
+    
+    const listaViagens = viagens.map((trip) => {   
+        const irParaDetalhes = (idViagem) => {
             setIdViagem(trip.id)
-            getTripDetail()
+            
             history.push(`/admin/trips/${idViagem}`)
-            return (
-                <div>
-                    <CardViagem key={trip.id}>
-                        <p><b>Nome: </b>{trip.name}</p>
-                        <p><b>Descrição: </b>{trip.description}</p>
-                        <p><b>Planeta: </b>{trip.planet}</p>
-                        <p><b>Duração: </b>{trip.durationInDays} dias</p>
-                        <p><b>Data: </b>{trip.date}</p>
-                    </CardViagem>
-                </div>
-            )
+            
         }
-
-        const getTripDetail = (idViagem) => {
-            const aluno = 'ruana-piber-carver'
-            const token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6ImlHUHVWYkJHMHVSTURab0JUZnFLIiwiZW1haWwiOiJydWFuYS5waWJlckBnbWFpbC5jb20iLCJpYXQiOjE2MzkwMDgwNzF9.wNM75qCArYR44HBAr2iUwge1lwEtqJ8mORLCmjgsyhY"
-            const headers = { auth: token }
-            const url = `https://us-central1-labenu-apis.cloudfunctions.net/labeX/${aluno}/trip/${idViagem}`
-            axios.get(url, headers).then((res) => {
-                setViagem(res.data.trip)
-                console.log('funciona')
-            }).catch((err) => {
-                console.log(err.data)
-            })
-        }
-
         return (
             <CardNomeViagem
                 key={idViagem}
